First pass splitting out things: - .NET split out - Added in babel (for esm modules, native Node modules are still experimental)
