+++ package.json
@@ -4,14 +4,25 @@
   "description": "OpenActive model generators",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "app-dev": "babel-node src/app.js",
+    "app": "node dist/app.js",
+    "build": "babel src --out-dir dist"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@babel/node": "^7.6.2",
+    "@babel/polyfill": "^7.6.0",
     "@openactive/data-models": "^1.3.7",
+    "babel-preset-env": "^1.7.0",
     "fs-extra": "^7.0.1",
     "sync-request": "^6.0.0"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.6.2",
+    "@babel/core": "^7.6.2",
+    "@babel/preset-env": "^7.6.2"
   }
 }
